requirements.txt
# -------------------------------
# Core Python Utilities
# -------------------------------
numpy>=1.23.0
pandas>=1.5.0
regex>=2023.12.25

# -------------------------------
# Web Framework (API)
# -------------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# -------------------------------
# Document & PDF Processing
# -------------------------------
PyMuPDF>=1.23.5          # fitz module
pdfplumber>=0.11.0
pytesseract>=0.3.10
pillow>=10.0.0

# -------------------------------
# OCR + Vision Extraction
# -------------------------------
paddleocr>=2.7.3
paddlepaddle>=2.6.1
opencv-python>=4.9.0.80

# -------------------------------
# Embedding & Vectorstore
# -------------------------------
chromadb>=0.5.0
sentence-transformers>=2.3.1
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.1.0

# -------------------------------
# Local LLM (Ollama)
# -------------------------------
ollama>=0.1.7

# -------------------------------
# Math & Finance Calculations
# -------------------------------
sympy>=1.12
scipy>=1.11.4

# -------------------------------
# Frontend / Cross-Origin
# -------------------------------
httpx>=0.27.0
aiohttp>=3.9.5

# -------------------------------
# Logging, Utils, and Dev Tools
# -------------------------------
tqdm>=4.66.1
rich>=13.7.0
python-dotenv>=1.0.1
